workspace/requirements-prod.txt
# MachineNativeOps Production Dependencies
# 生產環境依賴包

# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Cache and Message Queue
redis==5.0.1
aioredis==2.0.1
celery==5.3.4
kombu==5.3.4

# Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
cryptography==44.0.1

# Monitoring and Logging
prometheus-client==0.19.0
jaeger-client==4.8.0
structlog==23.2.0
python-json-logger==2.0.7

# HTTP client
httpx==0.25.2
aiohttp==3.13.3

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0

# Validation and Parsing
email-validator==2.1.0
python-dateutil==2.8.2
pytz==2023.3

# Business Logic
numpy==1.25.2
pandas==2.1.4
scipy==1.11.4

# File Processing
openpyxl==3.1.2
python-magic==0.4.27

# Development Tools (production safe)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==25.12.0
isort==5.12.0
flake8==6.1.0

# Performance
uvloop==0.19.0
orjson==3.11.5